remove data from DeepDAO
