+++ lib/profile/__tests__/getAggregatedData.spec.ts
@@ -2,17 +2,12 @@ import type { Space, SpaceRole } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
 import fetchMock from 'fetch-mock-jest';
 
-import { DEEPDAO_BASE_URL } from 'lib/deepdao/client';
 import { getAggregatedData } from 'lib/profile';
 import { randomETHWalletAddress } from 'lib/utilities/blockchain';
 import type { LoggedInUser } from 'models';
 import { generateBountyWithSingleApplication, generateUserAndSpaceWithApiToken } from 'testing/setupDatabase';
 
 const mockSandbox = fetchMock.sandbox();
-
-jest.mock('undici', () => {
-  return { fetch: (...args: any[]) => mockSandbox(...args) };
-});
 
 let user: LoggedInUser;
 let space: Space & { spaceRoles: SpaceRole[] };
@@ -45,91 +40,15 @@ describe('GET /api/public/profile/[userPath]', () => {
       userId: user.id
     });
 
-    const proposal1 = {
-      organizationId: '1',
-      createdAt: new Date().toString(),
-      title: undefined
-    };
-    const proposal2 = {
-      organizationId: '1',
-      createdAt: new Date().toString(),
-      title: undefined
-    };
-
-    const vote1 = {
-      organizationId: '1',
-      createdAt: new Date().toString()
-    };
-    const vote2 = {
-      organizationId: '1',
-      createdAt: new Date().toString()
-    };
-    const vote3 = {
-      organizationId: '2',
-      createdAt: new Date().toString()
-    };
-
-    mockSandbox
-      .get(`${DEEPDAO_BASE_URL}/v0.1/people/profile/${walletAddresses[0]}`, {
-        data: {
-          totalProposals: 1,
-          proposals: [proposal1, proposal2],
-          totalVotes: 1,
-          votes: [vote1, vote2],
-          organizations: [{ organizationId: '1', name: 'organization 1' }]
-        }
-      })
-      .get(`${DEEPDAO_BASE_URL}/v0.1/people/profile/${walletAddresses[1]}`, {
-        data: {
-          totalProposals: 2,
-          proposals: [],
-          totalVotes: 3,
-          votes: [vote3],
-          organizations: [{ organizationId: '2', name: 'organization 2' }]
-        }
-      })
-      .get(`${DEEPDAO_BASE_URL}/v0.1/organizations`, {
-        data: {
-          resources: [],
-          totalResources: 0
-        }
-      });
-
-    const aggregatedData = await getAggregatedData(user.id, 'dummy_key');
+    const aggregatedData = await getAggregatedData(user.id);
 
     expect(mockSandbox).toBeDone();
 
     expect(aggregatedData).toStrictEqual({
       bounties: 1,
-      totalProposals: 3,
-      totalVotes: 4,
+      totalProposals: 0,
+      totalVotes: 0,
       communities: [
-        {
-          id: '1',
-          name: 'organization 1',
-          isHidden: false,
-          isPinned: false,
-          logo: null,
-          joinDate: proposal1.createdAt,
-          votes: [vote1, vote2],
-          proposals: [proposal1, proposal2],
-          bounties: [],
-          latestEventDate: vote2.createdAt,
-          walletId: expect.any(String)
-        },
-        {
-          id: '2',
-          name: 'organization 2',
-          isPinned: false,
-          isHidden: false,
-          logo: null,
-          joinDate: vote3.createdAt,
-          votes: [vote3],
-          proposals: [],
-          bounties: [],
-          latestEventDate: vote3.createdAt,
-          walletId: expect.any(String)
-        },
         {
           id: space.id,
           joinDate: space.spaceRoles[0].createdAt.toISOString(),
