+++ pages/api/profile/orgs/[userId].ts
@@ -3,7 +3,7 @@ import nc from 'next-connect';
 
 import { onError, onNoMatch, requireUser } from 'lib/middleware';
 import type { UserCommunity } from 'lib/profile';
-import { getOrgs } from 'lib/profile/getOrgs';
+import { getCommunities } from 'lib/profile/getCommunities';
 import { withSessionRoute } from 'lib/session/withSession';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
@@ -13,9 +13,9 @@ handler.use(requireUser).get(getUserOrgs);
 async function getUserOrgs(req: NextApiRequest, res: NextApiResponse<UserCommunity[]>) {
   const { userId } = req.query as { userId: string };
 
-  const { charmverseCommunities, deepdaoCommunities } = await getOrgs({ userId });
+  const communities = await getCommunities({ userId });
 
-  return res.status(200).json([...charmverseCommunities, ...deepdaoCommunities]);
+  return res.status(200).json(communities);
 }
 
 export default withSessionRoute(handler);
